modularisation of the components in the `createEvent` page
